requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
Pillow>=10.4.0
Werkzeug>=3.0.0